requirements.txt
# Core dependencies
streamlit>=1.28.0
gradio>=4.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# AI/ML
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
diffusers>=0.24.0
accelerate>=0.25.0

# Audio
librosa>=0.10.0
soundfile>=0.12.0
pyttsx3>=2.90

# Video/Image
opencv-python>=4.8.0
pillow>=10.0.0
moviepy>=1.0.3

# Utilities
requests>=2.31.0
websockets>=12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
redis>=5.0.0

# Monitoring
psutil>=5.9.0
gpustat>=1.1.0
